+++ src/test/static/exported-behaviors.js
@@ -1,0 +1,11 @@
+/** @polymerBehavior */
+export const BasicBehavior1 = { method1() { }};
+
+/** @polymerBehavior */
+export const BasicBehavior2 = { method2() { } };
+
+/** @polymerBehavior */
+export const ArrayOfBehaviors = [BasicBehavior1, BasicBehavior2];
+
+/** @polymerBehavior */
+export default { method3() { } };