+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * [BREAKING] Import features will now be produced when a document has a broken
   import (e.g. because the imported file does not exist, or it does not parse,
   etc). We still emit a warning in such cases.
+* Recognize behaviors declared in export statements.
 <!-- Add new, unreleased changes here. -->
 
 ## [3.0.0-pre.17] - 2018-03-20